+++ src/main/java/app/model/beverage/Coffee.java
@@ -1,7 +1,7 @@
 package app.model.beverage;
 
 public class Coffee extends Beverage {
-    public Coffee(Float price) {
+    public Coffee(Integer price) {
         super(price);
     }
 }
